split tsconfig.json tsconfig.compile.json
